+++ packages/file-association/src/Exception/PropertyWriter/PropertyWriterException.php
@@ -15,7 +15,7 @@ namespace Rekalogika\File\Association\Exception\PropertyWriter;
 
 use Rekalogika\File\Association\Exception\FileAssociationException;
 
-final class PropertyWriterException extends FileAssociationException
+final class PropertyWriterException extends \LogicException implements FileAssociationException
 {
     /**
      * @param object|array<array-key,mixed> $object